+++ packages/forge-cli/src/cli/node/index.js
@@ -1,5 +1,6 @@
+exports.debug = require('./debug');
 exports.ps = require('./ps');
 exports.reboot = require('./reboot');
+exports.restart = require('./restart');
 exports.start = require('./start');
-exports.status = require('./status');
 exports.stop = require('./stop');