+++ packages/forge-cli/src/cli/node/ps/ps.js
@@ -1,58 +1,29 @@
-// const inquirer = require('inquirer');
 const shell = require('shelljs');
-const pidUsageTree = require('pidusage-tree');
-const pidInfo = require('find-process');
+const chalk = require('chalk');
 const Table = require('cli-table-redemption');
-const prettyTime = require('pretty-ms');
-const prettyBytes = require('pretty-bytes');
-const { config } = require('core/env');
-
-// const { symbols } = require('core/ui');
-// const { enums } = require('@arcblock/forge-proto');
+const { getForgeProcesses } = require('core/env');
+const { symbols } = require('core/ui');
 
 async function main() {
-  const { forgeBinPath, forgeConfigPath } = config.cli;
-  const { stdout: pid } = shell.exec(`FORGE_CONFIG=${forgeConfigPath} ${forgeBinPath} pid`, {
-    silent: true,
+  const processes = await getForgeProcesses();
+
+  if (!processes.length) {
+    shell.echo(
+      `${symbols.warning} forge daemon not started yet, start with ${chalk.cyan('forge start')}`
+    );
+    shell.echo('');
+    process.exit(0);
+  }
+
+  const table = new Table({
+    head: ['Name', 'PID', 'Uptime', 'Memory', 'CPU'],
+    style: { 'padding-left': 1, head: ['cyan', 'bold'], compact: true },
+    colWidths: [15, 10, 15, 15, 20],
   });
 
-  try {
-    const pidNumber = Number(pid);
-    const processes = await pidUsageTree(pidNumber);
-    const results = await Promise.all(
-      Object.values(processes).map(async x => {
-        try {
-          const [info] = await pidInfo('pid', x.pid);
-          Object.assign(x, info);
-        } catch (err) {
-          console.error(`Error getting pid info: ${x.pid}`, err);
-        }
-
-        return x;
-      })
-    );
-
-    const table = new Table({
-      head: ['Name', 'PID', 'Uptime', 'Memory', 'CPU'],
-      style: { 'padding-left': 1, head: ['cyan', 'bold'], compact: true },
-      colWidths: [15, 10, 15, 15, 20],
-    });
-    const rows = results
-      .filter(x => /(forge|tendermint|ipfs)/.test(x.name))
-      .map(x => ({
-        Name: x.name,
-        PID: x.pid,
-        Uptime: prettyTime(x.elapsed),
-        Memory: prettyBytes(x.memory),
-        CPU: x.cpu,
-      }));
-
-    rows.forEach(x => table.push(Object.values(x)));
-    console.log(table.toString());
-    process.exit(0);
-  } catch (err) {
-    console.error(err);
-  }
+  processes.forEach(x => table.push(Object.values(x)));
+  shell.echo(table.toString());
+  process.exit(0);
 }
 
 exports.run = main;