+++ packages/forge-cli/src/cli/tx/send/send.js
@@ -28,10 +28,10 @@ const questions = [
     type: 'editor',
     name: 'itx',
     message: 'Please enter the itx data object (js supported):',
-    default: answers => pretty(fakeMessage(answers.type)),
+    default: answers => pretty(fakeMessage(answers.type), { colors: false }),
     validate: x => {
       try {
-        safeEval(x);
+        safeEval(x, { client });
       } catch (err) {
         return err.message || err.toString();
       }
@@ -43,7 +43,7 @@ const questions = [
 
 async function main(data) {
   const { type, itx: itxStr } = data;
-  const itx = safeEval(itxStr);
+  const itx = safeEval(itxStr, { client });
   const { wallet } = config.cli;
 
   try {
@@ -63,6 +63,7 @@ async function main(data) {
 }
 
 function run() {
+  // console.log(Object.keys(client));
   inquirer.prompt(questions).then(main);
 }