+++ packages/forge-cli/src/cli/tx/sign/sign.js
@@ -27,10 +27,10 @@ const questions = [
     type: 'editor',
     name: 'itx',
     message: 'Please enter the inner transaction:',
-    default: answers => pretty(fakeMessage(answers.type)),
+    default: answers => pretty(fakeMessage(answers.type), { colors: false }),
     validate: x => {
       try {
-        safeEval(x);
+        safeEval(x, { client });
       } catch (err) {
         return err.message || err.toString();
       }
@@ -42,7 +42,7 @@ const questions = [
 
 function main(data) {
   const { type, itx: itxStr } = data;
-  const itx = safeEval(itxStr);
+  const itx = safeEval(itxStr, { client });
   const { wallet } = config.cli;
 
   return new Promise(resolve => {