+++ packages/forge-cli/src/core/ui.js
@@ -1,5 +1,6 @@
 const ora = require('ora');
 const util = require('util');
+const inquirer = require('inquirer');
 const symbols = require('log-symbols');
 
 // TODO: random spinner style
@@ -9,7 +10,9 @@ const symbols = require('log-symbols');
 
 module.exports = {
   symbols,
-  pretty: data => util.inspect(data, { depth: 8, colors: true, compact: false }),
+  hr: new inquirer.Separator().line,
+  pretty: (data, options) =>
+    util.inspect(data, Object.assign({ depth: 8, colors: true, compact: false }, options)),
   getSpinner: opts => {
     const spinner = ora(opts);
     return spinner;